Refactor landing view to use MVP
