+++ app/src/main/java/org/wit/hillfort/views/landing/LandingView.kt
@@ -1,28 +1,28 @@
-package org.wit.hillfort.activities
+package org.wit.hillfort.views.landing
 
 import android.os.Bundle
 import android.support.v7.app.AppCompatActivity
 import android.view.Menu
 import kotlinx.android.synthetic.main.activity_landing.*
-import org.jetbrains.anko.intentFor
 import org.wit.hillfort.R
-import org.wit.hillfort.views.login.LoginView
-import org.wit.hillfort.views.signup.SignupView
 
-class LandingActivity: AppCompatActivity() {
+class LandingView: AppCompatActivity() {
+  lateinit var presenter: LandingPresenter
 
   override fun onCreate(savedInstanceState: Bundle?) {
     super.onCreate(savedInstanceState)
     setContentView(R.layout.activity_landing)
 
+    presenter = LandingPresenter(this)
+
     // Start signup activity when button is pressed
     btnSignup.setOnClickListener{
-      startActivityForResult(intentFor<SignupView>(), 0)
+      presenter.doSignup()
     }
 
     // Start login activity when button is pressed
     btnLogin.setOnClickListener{
-      startActivityForResult(intentFor<LoginView>(), 0)
+      presenter.doLogin()
     }
   }
 
@@ -31,5 +31,4 @@ class LandingActivity: AppCompatActivity() {
     menuInflater.inflate(R.menu.menu_plain, menu)
     return super.onCreateOptionsMenu(menu)
   }
-
 }
